Extract reusable `FeedStoreSpecs` helper methods into a shared scope so that it can be used by other `FeedStore` implementation tests
